refactor(pipeline): move e2e test fixtures to proper project directory
Move inline fixture strings from PipelineTestProject.writeFixtureFiles()
to packages/integration-tests/src/test-projects/pipeline/, following the
established repo pattern (matching standalone-hosting-spa, hosting-test-app).

Static files (backend.ts, auth/, data/, hosting.ts, public/) now live as
real files on disk and are copied via fs.cp(). Only pipeline.ts remains
dynamically generated (it requires the runtime-generated stack name).

- Add test-projects/pipeline/ to integration-tests tsconfig exclude list
- All 71 unit tests passing, TypeScript and ESLint clean
