+++ packages/integration-tests/src/test-projects/pipeline/amplify/auth/resource.ts
@@ -1,0 +1,7 @@
+import { defineAuth } from '@aws-amplify/backend';
+
+export const auth = defineAuth({
+  loginWith: {
+    email: true,
+  },
+});
