+++ packages/integration-tests/src/test-projects/pipeline/amplify/backend.ts
@@ -1,0 +1,8 @@
+import { defineBackend } from '@aws-amplify/backend';
+import { auth } from './auth/resource.js';
+import { data } from './data/resource.js';
+
+defineBackend({
+  auth,
+  data,
+});
