+++ packages/integration-tests/src/test-projects/pipeline/amplify/hosting.ts
@@ -1,0 +1,6 @@
+import { defineHosting } from '@aws-amplify/hosting';
+
+defineHosting({
+  framework: 'spa',
+  buildOutputDir: 'public',
+});
